spam-function-source/requirements.txt
pandas 
scikit-learn 
xgboost 
gcsfs
google-cloud-storage 
joblib






